backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
reportlab==4.0.4
plotly==5.18.0
pandas==2.1.3
kaleido==0.2.1
